project deployed with java ee/maven/wildfly
