--- src/com/qa/account/accountapp/App.java
@@ -1,17 +1,0 @@
-package com.qa.account.accountapp;
-
-/**
- * Hello world!
- *
- */
-public class App
-
-
-{
-	private App() {}
-	
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
