+++ src/main/java/com/qa/account/accountapp/AccountInterface.java
@@ -1,0 +1,8 @@
+package com.qa.account.accountapp;
+
+public interface AccountInterface {
+
+	void addAccount(Account account);
+	
+	Account retrieveAccount(String accountNumber);
+}
